requirements_minimal.txt
# Minimal requirements for RAGbot - essentials only
# This avoids complex dependencies that might cause build issues

# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# Document processing (simplified)
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2

# Security
cryptography==41.0.7

# Database (using SQLite, included in Python)
# No additional requirements needed

# For ChromaDB (your existing vector store)
chromadb==0.4.22

# Basic ML (CPU only, smaller models)
sentence-transformers==2.2.2
# Note: torch will be installed as dependency

# Gradio (for existing UIs)
gradio==4.19.0